split TS file
