+++ ts/beamstreams.ts
@@ -19,108 +19,12 @@
   14 Ormiscaig, Aultbea, Achnasheen, IV22 2JJ  U.K.
 */
 import bindings from 'bindings';
-import { Frame, Stream, Muxer, CodecPar, Codec, Governor, WritableDemuxerStream, ReadableMuxerStream, FiltererVideoOptions, FiltererAudioOptions, Filterer, Decoder, Demuxer, Encoder, InputFormat } from '..'; // Codec, CodecContext, 
+import { Frame, Stream, Muxer, CodecPar } from '..'; // Codec, CodecContext, 
 import { Writable, Readable, Transform } from 'stream';
+import type { BeamcoderType, governorType, BeamstreamStream, BeamstreamSource } from './types';
 
-type governorType = {
-  read(len: number): Promise<Buffer>;
-  write(data: Buffer): Promise<null>;
-  finish(): undefined;
-};
-
-interface BeamcoderType extends ReadableMuxerStream {
-  /** Create object for AVIOContext based buffered I/O */
-  governor: typeof Governor;
-
-  /**
-   * Create a demuxer for this source
-   * @param options a DemuxerCreateOptions object
-   * @returns a promise that resolves to a Demuxer when it has determined sufficient 
-   * format details by consuming data from the source. The promise will wait indefinitely 
-   * until sufficient source data has been read.
-   */
-  // this code look to have error....
-  demuxer(options: { governor?: governorType, url?: string, iformat?: InputFormat, options?: { governor: governorType } } | string): Promise<Demuxer>
-  // url: src.url, iformat: src.iformat, options: src.options 
-  /**
-  * Provides a list and details of all the available encoders
-  * @returns an object with name and details of each of the available encoders
-  */
-  encoders(): { [key: string]: Codec };
-  /**
-   * Create a WritableDemuxerStream to allow streaming to a Demuxer
-   * @param options.highwaterMark Buffer level when `stream.write()` starts returng false.
-   * @returns A WritableDemuxerStream that can be streamed to.
-   */
-  demuxerStream(options: { highwaterMark?: number }): WritableDemuxerStream;
-
-  /**
-   * Create a ReadableMuxerStream to allow streaming from a Muxer
-   * @param options.highwaterMark The maximum number of bytes to store in the internal buffer before ceasing to read from the underlying resource.
-   * @returns A ReadableMuxerStream that can be streamed from.
-   */
-  muxerStream(options: { highwaterMark?: number }): ReadableMuxerStream;
-  /**
-   * Create a filterer
-   * @param options parameters to set up the type, inputs, outputs and spec of the filter
-   * @returns Promise that resolve to a Filterer on success
-   */
-  filterer(options: FiltererVideoOptions | FiltererAudioOptions): Promise<Filterer>
-
-  /** 
-   * Create a decoder by name
-   * @param name The codec name required
-   * @param ... Any non-readonly parameters from the Decoder object as required
-   * @returns A Decoder object - note creation is synchronous
-   */
-  decoder(options: { name: string, [key: string]: any }): Decoder
-  /**
-   * Create a decoder by codec_id
-   * @param codec_id The codec ID from AV_CODEC_ID_xxx
-   * @param ... Any non-readonly parameters from the Decoder object as required
-   * @returns A Decoder object - note creation is synchronous
-   */
-  decoder(options: { codec_id: number, [key: string]: any }): Decoder
-  /**
-   * Create a decoder from a demuxer and a stream_index
-   * @param demuxer An initialised Demuxer object
-   * @param stream_index The stream number of the demuxer object to be used to initialise the decoder
-   * @param ... Any non-readonly parameters from the Decoder object as required
-   * @returns A Decoder object - note creation is synchronous
-   */
-  decoder(options: { demuxer: Demuxer, stream_index: number, [key: string]: any }): Decoder
-  /**
-   * Create a decoder from a CodecPar object
-   * @param params CodecPar object whose codec name or id will be used to initialise the decoder
-   * @param ... Any non-readonly parameters from the Decoder object as required
-   * @returns A Decoder object - note creation is synchronous
-   */
-  decoder(options: { params: CodecPar, [key: string]: any }): Decoder
-  /**
-   * Create a frame for encoding or filtering
-   * Set parameters as required from the Frame object
-   */
-  frame(options: { [key: string]: any, data?: Array<Buffer> } | string): Frame;
-  /**
-   * Provides a list and details of all the available encoders
-   * @returns an object with name and details of each of the available encoders
-   */
-  encoders(): { [key: string]: Codec };
-  /** 
-   * Create an encoder by name
-   * @param name The codec name required
-   * @param ... Any non-readonly parameters from the Encoder object as required
-   * @returns An Encoder object - note creation is synchronous
-   */
-  encoder(options: { name: string, [key: string]: any }): Encoder
-  /**
-   * Create an encoder by codec_id
-   * @param codec_id The codec ID from AV_CODEC_ID_xxx
-   * @param ... Any non-readonly parameters from the Encoder object as required
-   * @returns An Encoder object - note creation is synchronous
-   */
-  encoder(options: { codec_id: number, [key: string]: any }): Encoder
-}
+import { teeBalancer } from './teeBalancer';
+import { parallelBalancer } from './parallelBalancer';
 
 const beamcoder = bindings('beamcoder') as BeamcoderType;
 
@@ -268,156 +172,6 @@ class serialBalancer {
 }
 
 
-type parallelBalancerType = Readable & {
-  pushPkts: (packets, stream, streamIndex: number, final?: boolean) => any
-};
-
-function parallelBalancer(params: { name: string, highWaterMark: number }, streamType: 'video' | 'audio', numStreams: number): parallelBalancerType {
-  let resolveGet = null;
-  const tag = 'video' === streamType ? 'v' : 'a';
-  const pending = [];
-  // initialise with negative ts and no pkt
-  // - there should be no output until each stream has sent its first packet
-  for (let s = 0; s < numStreams; ++s)
-    pending.push({ ts: -Number.MAX_VALUE, streamIndex: s });
-
-  const makeSet = resolve => {
-    if (resolve) {
-      // console.log('makeSet', pending.map(p => p.ts));
-      const nextPends = pending.every(pend => pend.pkt) ? pending : null;
-      const final = pending.filter(pend => true === pend.final);
-      if (nextPends) {
-        nextPends.forEach(pend => pend.resolve());
-        resolve({
-          value: nextPends.map(pend => {
-            return { name: `in${pend.streamIndex}:${tag}`, frames: [pend.pkt] };
-          }),
-          done: false
-        });
-        resolveGet = null;
-        pending.forEach(pend => Object.assign(pend, { pkt: null, ts: Number.MAX_VALUE }));
-      } else if (final.length > 0) {
-        final.forEach(f => f.resolve());
-        resolve({ done: true });
-      } else {
-        resolveGet = resolve;
-      }
-    }
-  };
-
-  const pushPkt = async (pkt, streamIndex: number, ts: number) =>
-    new Promise(resolve => {
-      Object.assign(pending[streamIndex], { pkt, ts, final: pkt ? false : true, resolve });
-      makeSet(resolveGet);
-    });
-
-  const pullSet = async () => new Promise<{ done: any, value: { timings: any } }>(resolve => makeSet(resolve));
-
-  const readStream: parallelBalancerType = new Readable({
-    objectMode: true,
-    highWaterMark: params.highWaterMark ? params.highWaterMark || 4 : 4,
-    read() {
-      (async () => {
-        const start = process.hrtime();
-        const reqTime = start[0] * 1e3 + start[1] / 1e6;
-        const result = await pullSet();
-        if (result.done)
-          this.push(null);
-        else {
-          result.value.timings = result.value[0].frames[0].timings;
-          result.value.timings[params.name] = { reqTime: reqTime, elapsed: process.hrtime(start)[1] / 1000000 };
-          this.push(result.value);
-        }
-      })();
-    },
-  }) as parallelBalancerType;
-
-  readStream.pushPkts = (packets: {frames: Array<any>, timings: [number, number]}, stream: {time_base: [number, number]}, streamIndex: number, final = false): any => {
-    if (packets && packets.frames.length) {
-      return packets.frames.reduce(async (promise, pkt) => {
-        await promise;
-        const ts = pkt.pts * stream.time_base[0] / stream.time_base[1];
-        pkt.timings = packets.timings;
-        return pushPkt(pkt, streamIndex, ts);
-      }, Promise.resolve());
-    } else if (final) {
-      return pushPkt(null, streamIndex, Number.MAX_VALUE);
-    }
-  };
-
-  return readStream;
-}
-
-type teeBalancerType = Readable[] & { pushFrames: (frames: any, unusedFlag?: boolean) => any };
-
-function teeBalancer(params: { name: 'streamTee', highWaterMark?: number }, numStreams: number): teeBalancerType {
-  let resolvePush = null;
-  const pending = [];
-  for (let s = 0; s < numStreams; ++s)
-    pending.push({ frames: null, resolve: null, final: false });
-
-  const pullFrame = async index => {
-    return new Promise<{ done: boolean, value?: any }>(resolve => {
-      if (pending[index].frames) {
-        resolve({ value: pending[index].frames, done: false });
-        Object.assign(pending[index], { frames: null, resolve: null });
-      } else if (pending[index].final)
-        resolve({ done: true });
-      else
-        pending[index].resolve = resolve;
-
-      if (resolvePush && pending.every(p => null === p.frames)) {
-        resolvePush();
-        resolvePush = null;
-      }
-    });
-  };
-
-  const readStreams: teeBalancerType = [] as teeBalancerType;
-  for (let s = 0; s < numStreams; ++s)
-    readStreams.push(new Readable({
-      objectMode: true,
-      highWaterMark: params.highWaterMark ? params.highWaterMark || 4 : 4,
-      read() {
-        (async () => {
-          const start = process.hrtime();
-          const reqTime = start[0] * 1e3 + start[1] / 1e6;
-          const result = await pullFrame(s);
-          if (result.done)
-            this.push(null);
-          else {
-            result.value.timings[params.name] = { reqTime: reqTime, elapsed: process.hrtime(start)[1] / 1000000 };
-            this.push(result.value);
-          }
-        })();
-      },
-    }));
-
-  readStreams.pushFrames = frames => {
-    return new Promise<{ value: { timings: any }, done: boolean }>(resolve => {
-      pending.forEach((p, index) => {
-        if (frames.length)
-          p.frames = frames[index].frames;
-        else
-          p.final = true;
-      });
-
-      pending.forEach(p => {
-        if (p.resolve) {
-          if (p.frames) {
-            p.frames.timings = frames.timings;
-            p.resolve({ value: p.frames, done: false });
-          } else if (p.final)
-            p.resolve({ done: true });
-        }
-        Object.assign(p, { frames: null, resolve: null });
-      });
-      resolvePush = resolve;
-    });
-  };
-
-  return readStreams;
-}
 
 function transformStream(params: { name: 'encode' | 'dice' | 'decode' | 'filter', highWaterMark: number }, processFn, flushFn, reject) {
   return new Transform({
@@ -686,25 +440,6 @@ function runStreams(
       streamTee[i].pipe(diceStream).pipe(encStream).pipe(muxStream);
     });
   });
-}
-
-export interface BeamstreamStream {
-  name: string;
-  time_base: any;
-  encoder: any;
-  stream: any;
-  codecpar: {
-    sample_rate: number;
-    frame_size: number;
-    format: any,
-    channel_layout: any;
-  };
-}
-export interface BeamstreamSource {
-  decoder: any;
-  format: any;
-  streamIndex: number;
-  stream: any;
 }
 
 export async function makeStreams(params: {
