+++ ts/teeBalancer.ts
@@ -1,0 +1,72 @@
+import { Readable } from "stream";
+
+type teeBalancerType = Readable[] & { pushFrames: (frames: any, unusedFlag?: boolean) => any };
+
+export function teeBalancer(params: { name: 'streamTee', highWaterMark?: number }, numStreams: number): teeBalancerType {
+  let resolvePush = null;
+  const pending = [];
+  for (let s = 0; s < numStreams; ++s)
+    pending.push({ frames: null, resolve: null, final: false });
+
+  const pullFrame = async index => {
+    return new Promise<{ done: boolean, value?: any }>(resolve => {
+      if (pending[index].frames) {
+        resolve({ value: pending[index].frames, done: false });
+        Object.assign(pending[index], { frames: null, resolve: null });
+      } else if (pending[index].final)
+        resolve({ done: true });
+      else
+        pending[index].resolve = resolve;
+
+      if (resolvePush && pending.every(p => null === p.frames)) {
+        resolvePush();
+        resolvePush = null;
+      }
+    });
+  };
+
+  const readStreams: teeBalancerType = [] as teeBalancerType;
+  for (let s = 0; s < numStreams; ++s)
+    readStreams.push(new Readable({
+      objectMode: true,
+      highWaterMark: params.highWaterMark ? params.highWaterMark || 4 : 4,
+      read() {
+        (async () => {
+          const start = process.hrtime();
+          const reqTime = start[0] * 1e3 + start[1] / 1e6;
+          const result = await pullFrame(s);
+          if (result.done)
+            this.push(null);
+          else {
+            result.value.timings[params.name] = { reqTime: reqTime, elapsed: process.hrtime(start)[1] / 1000000 };
+            this.push(result.value);
+          }
+        })();
+      },
+    }));
+
+  readStreams.pushFrames = frames => {
+    return new Promise<{ value: { timings: any }, done: boolean }>(resolve => {
+      pending.forEach((p, index) => {
+        if (frames.length)
+          p.frames = frames[index].frames;
+        else
+          p.final = true;
+      });
+
+      pending.forEach(p => {
+        if (p.resolve) {
+          if (p.frames) {
+            p.frames.timings = frames.timings;
+            p.resolve({ value: p.frames, done: false });
+          } else if (p.final)
+            p.resolve({ done: true });
+        }
+        Object.assign(p, { frames: null, resolve: null });
+      });
+      resolvePush = resolve;
+    });
+  };
+
+  return readStreams;
+}
